@@ -4,21 +4,22 @@ version: 0.0.0
 homepage: https://www.ademero.com
 
 environment:
-  sdk: '>=2.10.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   get_it: ^6.0.0
-  http: ^0.13.0
-  json_annotation: 3.1.1
-  json_serializable: 3.5.1
+  http: ^0.13.1
+  json_annotation: ^4.0.1
+  json_serializable: ^4.1.0
+  collection: ^1.15.0-nullsafety.4
 
 dev_dependencies:
-  build_runner: ^1.11.5
-  coverage: ^0.14.2
-  mockito: ^4.1.4
-  pedantic: ^1.9.0
+  build_runner: ^1.12.2
+  coverage: ^1.0.2
+  mockito: ^5.0.2
+  pedantic: ^1.11.0
   remove_from_coverage: ^1.0.7
-  test: ^1.15.7
+  test: ^1.16.8
 
 dependency_overrides:
   # Temporary to address the following development dependency error
@@ -26,3 +27,8 @@ dependency_overrides:
   #   And because coverage 1.0.2 depends on args ^2.0.0 and no versions of coverage match >1.0.2 <2.0.0, remove_from_coverage ^1.0
   # The remove_from_coverage package is only needed because the coverage package doesn't yet support ignoring generated (*.g.dart) files.
   args: ^2.0.0
+  # Temporary to address the following development dependency error
+  #   Because no versions of mockito match >5.0.2 <6.0.0 and mockito 5.0.2 depends on source_gen ^0.9.6, mockito ^5.0.2 requires source_gen ^0.9.6.
+  #   And because json_serializable >=4.1.0 depends on source_gen ^1.0.0, mockito ^5.0.2 is incompatible with json_serializable >=4.1.0.
+  # This should be resolved in the next mockito release, version >= 5.0.3.
+  source_gen: ^1.0.0
